requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-allauth==0.61.1
stripe==7.11.0
python-dotenv==1.0.1 